// apps/admin/src/lib/api/generated/models/userSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.7
 */
import type { AppearanceSettings } from './appearanceSettings';
import type { DefaultSettings } from './defaultSettings';
import type { NotificationSettings } from './notificationSettings';
import type { ProfileSettings } from './profileSettings';
import type { UserSettingsTables } from './userSettingsTables';

export interface UserSettings {
  appearance?: AppearanceSettings;
  defaults?: DefaultSettings;
  notifications?: NotificationSettings;
  profile?: ProfileSettings;
  tables?: UserSettingsTables;
}